Safer subprocess execution (#5)
- Replace unsafe os.system() with safer subprocess.run() calls
- Move more OS-specific subprocess invocations (unmountDisk(), eraseDisk(), validateNode()) into machdep_*.py
- Change Darwin implementation of privileged execution from osascript to sudo

Notably, this change removes the second confirmation prompt prior to erasing a disk in flashInstall.py (the command line invocation confirmation). This makes the default flow that much less annoying, at the expense of possibly allowing the erase command to run with unintended arguments if the user wasn't paying close attention. We may want to bring back the prompt.

Signed-off-by: pion <87685723+nbpion@users.noreply.github.com>